# doc/api.yaml
openapi: 3.0.0
paths:
...
  /user:
    get:
      operationId: UsersController_find
      summary: Gets all users
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Access token is missing or invalid
      tags:
        - User
    post:
      operationId: UsersController_create
      summary: Create user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserDto'
      responses:
        '201':
          description: The user has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request. body does not contain required fields
        '401':
          description: Access token is missing or invalid
      tags:
        - User
  /user/{id}:
    get:
      operationId: UsersController_findOne
      summary: Get single user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request. userId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: User not found
      tags:
        - User
    put:
      operationId: UsersController_update
      summary: Updates a user's password by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: The user has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request. userId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '403':
          description: oldPassowrd is wrong
        '404':
          description: User not found
      tags:
        - User
    delete:
      operationId: UsersController_delete
      summary: Deletes user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The user has been deleted
        '400':
          description: Bad request. userId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '403':
          description: oldPassowrd is wrong
        '404':
          description: User not found
      tags:
        - User
...
info:
  title: REST Service
  description: >-
    Let's try to create a Home Library Service! Users can create, read, update,
    delete data about Artists, Tracks and Albums, add them to Favorites in their
    own Home Library!
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    basic:
      type: http
      scheme: basic
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: d9683e62-ce01-4ac4-80c7-23828cd13792
          description: unique identifier
        login:
          type: string
          example: User
        version:
          type: number
          example: '1'
        createdAt:
          type: number
          example: 1675540182719
        updatedAt:
          type: number
          example: 1675540182719
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    AddUserDto:
      type: object
      properties:
        login:
          type: string
          default: Jimmy Roquai
        password:
          type: string
          default: '12345678'
      required:
        - login
        - password
    UpdateUserDto:
      type: object
      properties:
        oldPassword:
          type: string
          default: '12345678'
        newPassword:
          type: string
          default: '87654321'
      required:
        - oldPassword
        - newPassword
...